test1

+++ Partie 2/CRRPricer.h
@@ -1,7 +1,6 @@
 #pragma once
 #include <iostream>
 #include "BinaryTree.h"
-#include "BinaryTree.cpp"
 #include "Option.h"
 
 class CRRPricer {
@@ -18,9 +17,8 @@ public:
 	CRRPricer(Option*, int, double, double, double, double);
 	void compute();
 	double get(int n, int i) const;
-	int operator()(bool closed_form=false);
+	double operator()(bool closed_form=false);
 
 
 	~CRRPricer();
 };
-#include "CRRPricer.cpp"
